Split views.py into discrete controller files and functions
